Consider no tab to be selected while reconnect.
Also resolves that auto join channels are not reloaded on reconnect. As a side effect channels that are closed by user before reconnect(means explicit reconnect) are reopened, but I think that's fine, if user decides to reconnect(his fault XD). We could remember the autojoin channels the User left, but I thing reconnect should be a complete restart. "Consider no tab to be selected while reconnect" is happens because channels are removed on disconnect and no tab is selected so we need a null check there,
Fixes #806