Fix bugs in chat refactoring
Some bugs fixed in Downlord's chat rewrite. The categoriesToUserList map was never updated and no changes were made when the social status changed. Also all users were removed from one category in line 429 of ChannelTabController instead of only the items if the user that is updated. Further with new java version or javafx version the StringConverter of Combo boxes seems to be feed with null values as well(for empty display), so I returned "" in that case.

Fixes #976
Fixes #971