Fix two tabs being selected
JavaFX stores the index of the currently selected tab. When sorting the
tabs, this index isn't updated. When selecting a new tab, the preivously
selected tab gets deselected based on the currently selected tab index -
which is no longer the correct index. This causes the currently selected
tab to not be unselected, but the new one to be selected, too.

Fixes #998