Change "Ø Rating" to "Avg. Rating"
Change "Ø Rating" to "Avg. Rating" on the game list because this symbol is not commonly known to mean "Average" in some English-speaking countries. Should probably change in other localizations as well but I don't know how other languages would commonly abbreviate or otherwise represent "Average" so I don't want to change it myself and make it worse.