Implement auto-update
Uses update4j to update the client automatically. During the build
process, an update4j.xml is generated as well as an archive file that
contains all the required client files. Both are uploaded to a server
via SSH, where the archive is extracted.

The client will check if a new version is available, either using the
managed configuration file that resides on the FAF server (for stable
releases) or (for pre-releases) via the GitHub releases API.

If a new version is available, the client uses update4j to download all
files to a temporary location. Afterwards, a separate Java application
is spawned which waits for the client to terminate. Once that happens,
the application will finalize the update by copying the update files
over the existing installation and removing all obsolete files.

On next client start, the application is updated.

Fixes #1753