Use parseMapName over parseMapFolderName
For some reason parseMapFolderName returns the map name without any capitalization from the raw replay bytes. This causes issues with the base64 string encoding used by the next gen map generator. parseMapName returns the map name with correct capitalization making it more suitable for usage. As an aside it appears that parseMapFolderName uses the scenario file in the raw bytes to get the map name whereas parseMapName uses the .scmap file and the capitalization differences are reflected there.