For Linux this launches proton using steam ID 9420 to launch FA.
--

there is OS profile information in main with no obvious access other than the static methods written to run the bridj library calls.

instead of calling the bridj API I stuck to the PROFILE_STRING convention written.

I also noted there is no code setting the executableDecorator and it is unclear if there is DI that intellij is not hinting for me.  I took no liberies to bridge the unreachable PROFILE strings with the single-constant-instance of the decorator in the hopes that I can be shown any hierarchies I have missed with this patch.